Move RtpHeaderExtensionCapability into separate module
